Methodology edits
new make_tree function does not require reverse stack anymore-code is easier to read and functions recursively. __str__ now does not rely on secondary recursive call. Edited insert child functions-no longer creates a new node is the child is not a node. ExpTree static methods updated to account for this change-base cases of recursive's now check if not instance of ExpTree instead of checking for is None.